@@ -305,10 +305,10 @@ EVENTS: dict[int, dict[str, Alert | AlertCallbackType]] = {
 
   EventName.stockLkas: {
     ET.PERMANENT: Alert(
-      "TAKE CONTROL",
       "Stock LKAS: Lane Departure Detected",
-      AlertStatus.critical, AlertSize.full,
-      Priority.HIGH, VisualAlert.fcw, AudibleAlert.none, 2.),
+      "",
+      AlertStatus.userPrompt, AlertSize.small,
+      Priority.LOW, VisualAlert.ldw, AudibleAlert.prompt, 3.),
     ET.NO_ENTRY: NoEntryAlert("Stock LKAS: Lane Departure Detected"),
   },
 
