@@ -24,7 +24,7 @@ class PairingDialog(NavWidget):
     self._qr_texture: rl.Texture | None = None
     self._last_qr_generation = float("-inf")
 
-    self._txt_pair = gui_app.texture("icons_mici/settings/device/pair.png", 84, 64)
+    self._txt_pair = gui_app.texture("icons_mici/settings/device/pair.png", 33, 60)
     self._pair_label = MiciLabel("pair with comma connect", 48, font_weight=FontWeight.BOLD,
                                  color=rl.Color(255, 255, 255, int(255 * 0.9)), line_height=40, wrap_text=True)
 
